move type alias tests to the reflex_base suite
